feat(starter): switch the default Babel preset to use babel-preset-env
This PR switches the default Babel preset to use babel-preset-env - this
preset adjusts based on the current Electron version which Babel plugins
to use. This PR also enables async/await and React by default, since
these features are super common and don't have any overhead to enable
when not in-use.