Replace supportedPlatforms with supportsPlatform
Having each maker export an array of supported platforms ends up being non-deterministic because any platform not supported by a given maker will be unable to import and read its package.json's os field. So it really only makes sense to query if the maker is supported on the current platform, and that's really all we care about anyway.